pyproject.toml
[tool.poetry]
name = "metamove"
version = "0.1.0"
description = "Transform YAML: move meta under config.meta"
authors = ["Oliver Laslett <oliver@lightdash.com>"]
readme = "README.md"
packages = [{include = "metamove"}]
license = "MIT"
repository = "https://github.com/lightdash/metamove"
keywords = ["yaml", "transform", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.14"
ruamel-yaml = "0.18.6"
click = "^8.0.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pyinstaller = "^6.3.0"

[tool.poetry.scripts]
metamove = "metamove.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 